+++ marrow/src/_with_latest_arrow/test/mod.rs
@@ -1,2 +1,0 @@
-mod arrays;
-mod data_types;